data-consumer/requirements.txt
pymongo==4.3.3
kink==0.6.5
paho-mqtt==1.6.1
